requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorchvideo>=0.1.5
av>=10.0.0
numpy>=1.24.0
Pillow>=9.5.0
pyyaml>=6.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.2.0
streamlit>=1.28.0
opencv-python-headless>=4.8.0
tensorboard>=2.13.0
